✨ : 프로필 정보 api 분리
