+++ src/pages/profile/api.ts
@@ -1,16 +1,38 @@
+import { GetServerSidePropsContext } from 'next';
 import instance from '@/api/axios';
+import { getFeedList } from '@/components/Feed/FeedList/api';
+import { MemberDataType } from './types';
+import { FeedListType } from '@/components/Feed/types';
 
-export async function getMyProfile() {
+export async function fetchMemberData(context: GetServerSidePropsContext) {
+  const { req } = context;
+  const cookies = req.headers.cookie || '';
+  const { memberId } = context.query;
+  const endpoint = memberId ? `profile/${memberId}` : '/profile/mine';
+
   try {
-    const memberData = await instance.get('/profile/mine', {
-      // headers: {
-      //   'Authorization' : ``
-      // }
+    const res = await instance.get(endpoint, {
+      headers: {
+        Cookie: cookies,
+      },
     });
-    console.log(memberData);
-    return memberData;
+    const memberData: MemberDataType = await res.data;
+    const feedList: FeedListType = await getFeedList();
+
+    return {
+      props: {
+        feedList: feedList.data,
+        memberData,
+      },
+    };
   } catch (error) {
-    console.error('프로필 정보 가져오기 실패:', error);
-    throw error;
+    console.error('API 호출 실패ㅠ', error);
+    return {
+      props: {
+        feedList: [],
+        memberData: null,
+        error: true,
+      },
+    };
   }
 }
