+++ src/pages/profile/types.ts
@@ -1,0 +1,10 @@
+export interface MemberDataType {
+  memberId?: number;
+  nickname: string;
+  generation: number;
+  profileImageUrl: string;
+  introduce: string;
+  followerCount: number;
+  followingCount: number;
+  isAuthorized: boolean;
+}
